requirements.txt
# Core dependencies
ccxt==4.3.22
pandas==2.2.0
numpy==1.26.4
python-dotenv==1.0.0

# Technical analysis
ta==0.11.0

# Visualization
matplotlib==3.8.0
plotly==5.18.0

# Backtesting
backtesting==0.6.5

# Utilities
python-dateutil==2.9.0
pytz==2025.2
seaborn==0.13.2
schedule==1.2.0python-binance>=3.0.0
